// movie-store-web-app/src/app/api/api-reference.ts
export enum LicensingType {
    TwoDay = 1,
    Lifelong = 2,
}

export abstract class Movie implements IMovie {
    id?: string;
    name!: string;
    licensingType!: LicensingType;

    constructor(data?: IMovie) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.licensingType = _data["licensingType"];
        }
    }

    static fromJS(data: any): Movie {
        data = typeof data === 'object' ? data : {};
        throw new Error("The abstract class 'Movie' cannot be instantiated.");
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["licensingType"] = this.licensingType;
        return data;
    }
}

export interface IMovie {
    id?: string;
    name: string;
    licensingType: LicensingType;
}
// ...
export class CreateMovieCommand implements ICreateMovieCommand {
    name?: string;
    licensingType?: LicensingType;

    constructor(data?: ICreateMovieCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.licensingType = _data["licensingType"];
        }
    }

    static fromJS(data: any): CreateMovieCommand {
        data = typeof data === 'object' ? data : {};
        let result = new CreateMovieCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["licensingType"] = this.licensingType;
        return data;
    }
}

export interface ICreateMovieCommand {
    name?: string;
    licensingType?: LicensingType;
}

export class UpdateMovieCommand implements IUpdateMovieCommand {
    movieId?: string;
    name?: string;
    licensingType?: LicensingType;

    constructor(data?: IUpdateMovieCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.movieId = _data["movieId"];
            this.name = _data["name"];
            this.licensingType = _data["licensingType"];
        }
    }

    static fromJS(data: any): UpdateMovieCommand {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateMovieCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["movieId"] = this.movieId;
        data["name"] = this.name;
        data["licensingType"] = this.licensingType;
        return data;
    }
}

export interface IUpdateMovieCommand {
    movieId?: string;
    name?: string;
    licensingType?: LicensingType;
}